requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
dotenv==0.9.9
durationpy==0.9
Flask==3.1.0
google-auth==2.40.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.1.5
kubernetes==32.0.1
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
oauthlib==3.2.2
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pymongo==4.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
